#77 Added per-check field ingestion metrics
This will allow us to see the ratio of fields generated by checks and what can actually be written to Kairos as metrics.
Doing this required extracting and refactoring the code that transforms a JsonNode tree into a flat map of numerics.  The only location where all the data required for these metrics lives at the same time is in the WorkResultWriters, one of which already exists for writing per-account metrics.

It is questionable whether or not these metrics should be done with a parallel metrics writer or inline with the existing Kairos writer.  Doing this as a separate writer separates concerns and garantees continuing metrics in the event of a Kairos outage, however it also requires doubling the overall complexity as the same scanning and flattening has to be done in its entirety a second time.

Added and updated relevant tests.